Moving the adduser function to appropriate directory and modifying it for use with promises
